+++ src/test/java/com/jnape/palatable/lambda/adt/EitherTest.java
@@ -15,7 +15,6 @@ import testsupport.traits.MonadLaws;
 import testsupport.traits.TraversableLaws;
 
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
 
 import static com.jnape.palatable.lambda.adt.Either.fromMaybe;
 import static com.jnape.palatable.lambda.adt.Either.left;
@@ -23,7 +22,6 @@ import static com.jnape.palatable.lambda.adt.Either.right;
 import static com.jnape.palatable.lambda.adt.Maybe.just;
 import static com.jnape.palatable.lambda.adt.Maybe.nothing;
 import static com.jnape.palatable.lambda.adt.Unit.UNIT;
-import static com.jnape.palatable.lambda.functions.Effect.fromConsumer;
 import static com.jnape.palatable.lambda.functor.builtin.Lazy.lazy;
 import static com.jnape.palatable.traitor.framework.Subjects.subjects;
 import static org.hamcrest.core.Is.is;
@@ -199,37 +197,6 @@ public class EitherTest {
         assertEquals(right(UNIT), Either.trying(() -> {}));
         IllegalStateException expected = new IllegalStateException("expected");
         assertEquals(left(expected), Either.trying(() -> {throw expected;}));
-    }
-
-    @Test
-    public void monadicPeekLiftsIOToTheRight() {
-        Either<String, Integer> left  = left("foo");
-        Either<String, Integer> right = right(1);
-
-        AtomicInteger intRef = new AtomicInteger();
-
-        left.peek(fromConsumer(intRef::set));
-        assertEquals(0, intRef.get());
-
-        right.peek(fromConsumer(intRef::set));
-        assertEquals(1, intRef.get());
-    }
-
-    @Test
-    public void dyadicPeekDuallyLiftsIO() {
-        Either<String, Integer> left  = left("foo");
-        Either<String, Integer> right = right(1);
-
-        AtomicReference<String> stringRef = new AtomicReference<>();
-        AtomicInteger           intRef    = new AtomicInteger();
-
-        left.peek(fromConsumer(stringRef::set), fromConsumer(intRef::set));
-        assertEquals("foo", stringRef.get());
-        assertEquals(0, intRef.get());
-
-        right.peek(fromConsumer(stringRef::set), fromConsumer(intRef::set));
-        assertEquals("foo", stringRef.get());
-        assertEquals(1, intRef.get());
     }
 
     @Test